requirements.txt
schedule==1.2.2
openai
python-dotenv
